+++ src/Module/RenderCartBookingInfoHandler.php
@@ -4,6 +4,7 @@ namespace RebelCode\EddBookings\Cart\Module;
 
 use ArrayAccess;
 use Carbon\Carbon;
+use Dhii\Cache\ContainerInterface as CacheContainerInterface;
 use Dhii\Data\Container\ContainerGetCapableTrait;
 use Dhii\Data\Container\ContainerGetPathCapableTrait;
 use Dhii\Data\Container\ContainerHasCapableTrait;
@@ -14,6 +15,7 @@ use Dhii\Data\Container\NormalizeContainerCapableTrait;
 use Dhii\Data\Container\NormalizeKeyCapableTrait;
 use Dhii\Exception\CreateInvalidArgumentExceptionCapableTrait;
 use Dhii\Exception\CreateOutOfRangeExceptionCapableTrait;
+use Dhii\Exception\CreateRuntimeExceptionCapableTrait;
 use Dhii\I18n\StringTranslatingTrait;
 use Dhii\Invocation\InvocableInterface;
 use Dhii\Iterator\CountIterableCapableTrait;
@@ -21,12 +23,15 @@ use Dhii\Iterator\ResolveIteratorCapableTrait;
 use Dhii\Output\TemplateAwareTrait;
 use Dhii\Output\TemplateInterface;
 use Dhii\Storage\Resource\SelectCapableInterface;
+use Dhii\Util\Normalization\NormalizeArrayCapableTrait;
 use Dhii\Util\Normalization\NormalizeIntCapableTrait;
 use Dhii\Util\Normalization\NormalizeIterableCapableTrait;
 use Dhii\Util\Normalization\NormalizeStringCapableTrait;
 use Dhii\Util\String\StringableInterface as Stringable;
 use Psr\Container\ContainerInterface;
 use Psr\EventManager\EventInterface;
+use RebelCode\Bookings\BookingFactoryInterface;
+use RebelCode\Entity\GetCapableManagerInterface;
 use stdClass;
 
 /**
@@ -41,6 +46,12 @@ class RenderCartBookingInfoHandler implements InvocableInterface
 
     /* @since [*next-version*] */
     use GetBookingDisplayTimezoneCapableTrait;
+
+    /* @since [*next-version*] */
+    use GetBookingSessionTypeCapableTrait;
+
+    /* @since [*next-version*] */
+    use GetBookingSessionInfoCapableTrait;
 
     /* @since [*next-version*] */
     use ContainerGetPathCapableTrait;
@@ -67,6 +78,9 @@ class RenderCartBookingInfoHandler implements InvocableInterface
     use NormalizeStringCapableTrait;
 
     /* @since [*next-version*] */
+    use NormalizeArrayCapableTrait;
+
+    /* @since [*next-version*] */
     use NormalizeIterableCapableTrait;
 
     /* @since [*next-version*] */
@@ -88,6 +102,9 @@ class RenderCartBookingInfoHandler implements InvocableInterface
     use CreateInvalidArgumentExceptionCapableTrait;
 
     /* @since [*next-version*] */
+    use CreateRuntimeExceptionCapableTrait;
+
+    /* @since [*next-version*] */
     use StringTranslatingTrait;
 
     /**
@@ -98,6 +115,51 @@ class RenderCartBookingInfoHandler implements InvocableInterface
      * @var SelectCapableInterface
      */
     protected $bookingsSelectRm;
+
+    /**
+     * The factory for creating booking instances.
+     *
+     * @since [*next-version*]
+     *
+     * @var BookingFactoryInterface
+     */
+    protected $bookingFactory;
+
+    /**
+     * The services manager for retrieving services by ID.
+     *
+     * @since [*next-version*]
+     *
+     * @var GetCapableManagerInterface
+     */
+    protected $servicesManager;
+
+    /**
+     * The resources manager for retrieving resources by ID.
+     *
+     * @since [*next-version*]
+     *
+     * @var GetCapableManagerInterface
+     */
+    protected $resourcesManager;
+
+    /**
+     * The services cache.
+     *
+     * @since [*next-version*]
+     *
+     * @var CacheContainerInterface
+     */
+    protected $servicesCache;
+
+    /**
+     * The resources cache.
+     *
+     * @since [*next-version*]
+     *
+     * @var CacheContainerInterface
+     */
+    protected $resourcesCache;
 
     /**
      * The expression builder.
@@ -124,6 +186,11 @@ class RenderCartBookingInfoHandler implements InvocableInterface
      *
      * @param TemplateInterface                             $template         The template to use to render the info.
      * @param SelectCapableInterface                        $bookingsSelectRm The bookings SELECT resource model.
+     * @param BookingFactoryInterface                       $bookingFactory   The bookings factory.
+     * @param GetCapableManagerInterface                    $servicesManager  The services manager.
+     * @param GetCapableManagerInterface                    $resourcesManager The resources manager.
+     * @param CacheContainerInterface                       $serviceCache     The cache for service.
+     * @param CacheContainerInterface                       $resourceCache    The cache for resource.
      * @param object                                        $exprBuilder      The expression builder.
      * @param array|stdClass|ArrayAccess|ContainerInterface $cartItemConfig   The cart item data config.
      * @param string|Stringable|null                        $fallbackTz       The fallback timezone to use for bookings
@@ -132,15 +199,26 @@ class RenderCartBookingInfoHandler implements InvocableInterface
     public function __construct(
         TemplateInterface $template,
         SelectCapableInterface $bookingsSelectRm,
+        BookingFactoryInterface $bookingFactory,
+        GetCapableManagerInterface $servicesManager,
+        GetCapableManagerInterface $resourcesManager,
+        CacheContainerInterface $serviceCache,
+        CacheContainerInterface $resourceCache,
         $exprBuilder,
         $cartItemConfig,
         $fallbackTz
     ) {
         $this->_setTemplate($template);
+        $this->_setFallbackTz($fallbackTz);
+
         $this->bookingsSelectRm = $bookingsSelectRm;
+        $this->bookingFactory   = $bookingFactory;
+        $this->servicesManager  = $servicesManager;
+        $this->resourcesManager = $resourcesManager;
+        $this->servicesCache    = $serviceCache;
+        $this->resourcesCache   = $resourceCache;
         $this->exprBuilder      = $exprBuilder;
         $this->cartItemConfig   = $cartItemConfig;
-        $this->_setFallbackTz($fallbackTz);
     }
 
     /**
@@ -201,6 +279,10 @@ class RenderCartBookingInfoHandler implements InvocableInterface
      */
     protected function _renderBookingInfo($bookingData)
     {
+        $booking = $this->bookingFactory->make([
+            BookingFactoryInterface::K_DATA => $bookingData,
+        ]);
+
         $format   = $this->_containerGet($this->cartItemConfig, 'booking_datetime_format');
         $clientTz = $this->_getDisplayTimezone($bookingData);
 
@@ -222,15 +304,83 @@ class RenderCartBookingInfoHandler implements InvocableInterface
         $startStr = $startDt->format($format);
         $endStr   = $endDt->format($format);
 
+        // Get the matching session's info
+        $sessionInfo = $this->_getBookingSessionInfo($booking);
+        // Append the session label to the service name if it exists
+        $sessionLabel = $this->_containerGet($sessionInfo, 'session_label');
+
+        // Prepare the resources text
+        $resources     = $this->_containerGet($sessionInfo, 'resource_names');
+        $resourcesText = !empty($resources)
+            ? sprintf('%s %s', $this->__('with'), implode(', ', $resources))
+            : '';
+
         return $this->_getTemplate()->render([
-            'from_label'     => $this->__('From:'),
-            'until_label'    => $this->__('Until:'),
-            'timezone_label' => $this->__('Timezone:'),
-            'start_datetime' => $startTs,
-            'end_datetime'   => $endTs,
-            'start_text'     => $startStr,
-            'end_text'       => $endStr,
-            'timezone_text'  => $this->_normalizeTimezoneName($clientTz->getName())
+            'session_label'     => $sessionLabel,
+            'resources'         => $resourcesText,
+            'start_datetime'    => $startTs,
+            'end_datetime'      => $endTs,
+            'start_text'        => $startStr,
+            'end_text'          => $endStr,
+            'timezone_text'     => $this->_normalizeTimezoneName($clientTz->getName()),
         ]);
+    }
+
+    /**
+     * Retrieves the name of the service.
+     *
+     * @since [*next-version*]
+     *
+     * @param int|string|Stringable $serviceId The ID of the service.
+     *
+     * @return string|Stringable The name of the service.
+     */
+    protected function _getServiceName($serviceId)
+    {
+        $service = $this->servicesCache->get($serviceId, function ($serviceId) {
+            return $this->servicesManager->get($serviceId);
+        });
+
+        return $this->_containerGet($service, 'name');
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _getResourcesManager()
+    {
+        return $this->resourcesManager;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _getResourceCache()
+    {
+        return $this->resourcesCache;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _getServicesManager()
+    {
+        return $this->servicesManager;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @since [*next-version*]
+     */
+    protected function _getServiceCache()
+    {
+        return $this->servicesCache;
     }
 }